requirements.txt
flask
pyodbc
python-dotenv
requests
flask-cors
